+++ src/main/MfTable/getMessageByDiff.ts
