+++ src/main/manipulateBrowser/backDisplayMonthToTarget/generateDateLabel.ts
